Refactor recent service tracking implementation and add search parameter hook
