+++ apps/client/index.html
@@ -1,22 +1,21 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <title>FestifyWS</title>
-    <base href="/" />
 
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link
-      rel="icon"
-      type="image/x-icon"
-      href="favicon.ico"
-      data-react-helmet="true"
-    />
-    <link rel="stylesheet" href="/src/styles.css" />
-  </head>
+<head>
+  <meta charset="utf-8" />
+  <title>Festify Web Services</title>
+  <meta name="description"
+    content="The Festify Web Services is a set of services developers can use for their projects." />
+  <base href="/" />
 
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
-  </body>
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <link rel="icon" type="image/x-icon" href="favicon.ico" data-react-helmet="true" />
+  <link rel="stylesheet" href="/src/styles.css" />
+</head>
+
+<body>
+  <div id="root"></div>
+  <script type="module" src="/src/main.tsx"></script>
+</body>
+
 </html>
