+++ apps/client/packages/sharedui/src/hooks/useSearchParam.tsx
@@ -1,0 +1,9 @@
+import { useLocation } from 'react-router-dom';
+
+const useSearchParam = (key: string) => {
+  const location = useLocation();
+
+  return new URLSearchParams(location.search).get(key) || undefined;
+};
+
+export default useSearchParam;
